added authentication

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\BlogPosts;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(BlogPosts::all());
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        return PostResource::collection($posts);
     }
 
     /**
@@ -23,11 +26,10 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
-        $data= $request -> validated();   
-        $data["author_id"] = 1 ;
-       $blogPost = BlogPosts::create($data);
-        
-        return response()-> json($blogPost, 201);
+        $data = $request->validated();
+        $data['author_id'] = $request->user()->id;
+        $blogPost = BlogPosts::create($data);
+        return response()->json(new PostResource($blogPost), 201);
     }
 
     /**
@@ -35,15 +37,25 @@ class PostController extends Controller
      */
     public function show(Blogposts $post)
     {
-        return response()->json($post);
+        abort_if(Auth::id() !== $post->author_id, 403, 'Unauthorized');
+        return response()->json(new PostResource($post));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, BlogPosts $post)
     {
-        $request -> validated();
+
+        abort_if(Auth::id() !== $post->author_id, 403, 'Unauthorized');
+        $data = $request->validate(
+            [
+                'title' => 'sometimes|required|string|max:255',
+                'content' => 'sometimes|required|string',
+            ]
+        );
+        $post->update($data);
+        return response()->json(new PostResource($post));
     }
 
     /**
@@ -51,6 +63,7 @@ class PostController extends Controller
      */
     public function destroy(BlogPosts $post)
     {
+        abort_if(Auth::id() !== $post->author_id, 403, 'Unauthorized');
         $post->delete();
         return response()->noContent();
     }
